refactored DeadlineTracker.java into resources folder
